requirements.txt
black==21.5b0
coverage==5.5
flake8==3.9.1
isort==5.9.1
numpy==1.18.1
python-dotenv==0.19.0
requests==2.24.0
pandas==1.0.2
pandas-usaddress==0.21
setuptools==54.2.0